requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
alpaca-trade-api>=3.2.0
chromadb>=0.4.22
python-dotenv>=1.0.0
pytest>=8.0.0
pydantic>=2.5.0
requests>=2.31.0
